+++ CHANGELOG.md
@@ -93,3 +93,4 @@ Updates:
 * Code refactored to support next-gen JS engine
 * More short-cuts added to explorer (including multi-select)
 * Creation of new intent-types using templates
+* Deletion of intent-types with intents (confirmation prompt)